Backend/requirements.txt
fastapi                
google-generativeai    
grpcio-status          
langchain              
langchain-google-genai 
load-dotenv            
python-multipart       
uvicorn                